backend/requirements.txt
absl-py==2.2.2
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi
fonttools==4.57.0
google-auth==2.34.0
grpcio==1.71.0
httpx==0.27.2
idna==3.10
ipython==8.35.0
jedi==0.19.2
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
numpy==1.23.5
opencv-python==4.11.0.86
opencv-python-headless==4.8.1.78
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==4.21.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-jose[cryptography]
python-multipart
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sqlmodel
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn[standard]
watchfiles
wcwidth==0.2.13
Werkzeug==3.1.3
